Use non-alertable waits only for WM_PAINT handling
